Respond to review comments.
compiler/hlds_out_util.m:
compiler/llds_out_data.m:
compiler/llds_out_util.m:
    As above.

library/int64.m:
   Fix spelling.

library/uint64.m:
   Fix a cut-and-paste error.

+++ library/uint64.m
@@ -193,7 +193,7 @@
 "
     if (I < 0) {
         SUCCESS_INDICATOR = MR_FALSE;
-    } else if ((uint64_t)I > (uint64_t)UINT32_MAX) {
+    } else if ((uint64_t)I > (uint64_t)INT64_MAX) {
         SUCCESS_INDICATOR = MR_FALSE;
     } else {
         U = (uint64_t) I;
